requirements.txt
python-dotenv
country_converter~=1.3.1
pycountry~=24.6.1
peewee~=3.18.2
countryinfo~=0.1.2
numpy~=2.3.3
requests~=2.32.5
yt-dlp~=2025.9.5
websockets~=15.0.1
pytchat~=0.5.5
